Modified: ModuleResults now show the Duration since it was initialized. Modified: An error in a module now return the inner exception. Modified: Updated MCM and Geo Modules and extension files

+++ src/app/Geckon.Portal.Data/Result/IModuleResult.cs
@@ -5,7 +5,7 @@ namespace Geckon.Portal.Data.Result
     public interface IModuleResult
     {
         string Fullname { get; }
-        double Duration { get; }
+        long Duration { get; }
         int Count { get; }
         IList<IResult> Results { get; set; }
         void AddResult( IEnumerable<IResult> results );
